// src/app/utilities/models/bug.ts
export class Bug {
  'bugTitle': string;
  'Description': string;
  'hostAppVersion': string;
  'isSignedRelease': string;
  'deviceOS': string;
  'deviceModel': string;
  'tags': string;
  'videoAttachments': string;
  'audioAttachments': string;
  'images': string;
  'reporteeName': string;
  'reporteeEmail': string;
  'deviceId': string;
  'appPackageId': string;
  'isEmulator': string;
  'isRooted': string;
  'sdkVersionId': string;
  'userActivityLogs': string;
  'freeMemory': string;
  'freeDisk': string;
  'isOnBattery': string;
  'batteryLevel': string;
  'appPermissions': string;
  'assigneeName': string;
  'email': string;
  'assigneeKuid': string;
}